# perf: use insertion sort algorithm